Move Competitor entity into CompetitorAggregate
