+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/CompetitorAggregate/CompetitorTests.cs
@@ -1,10 +1,11 @@
 ﻿using FluentAssertions;
 using System;
 using System.Collections.Generic;
+using TournamentManagement.Domain.CompetitorAggregate;
 using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
-namespace TournamentManagement.Domain.UnitTests
+namespace TournamentManagement.Domain.UnitTests.CompetitorAggregate
 {
 	public class CompetitorTests
 	{
