+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/MatchAggregate/MatchTests.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using System;
 using System.Collections.Generic;
+using TournamentManagement.Domain.CompetitorAggregate;
 using TournamentManagement.Domain.MatchAggregate;
 using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
