+++ src/TournamentManagement/TournamentManagement.Domain/MatchAggregate/Match.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using TournamentManagement.Domain.CompetitorAggregate;
 using TournamentManagement.Domain.TournamentAggregate;
 
 namespace TournamentManagement.Domain.MatchAggregate
